feat : Comments controller & Dto

+++ back/src/main/java/cafemate/back/repository/CommentsRepository.java
@@ -4,6 +4,6 @@ import cafemate.back.domain.Comments;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface CommentsRepository extends JpaRepository<Comments, Long> {
-    List<Comments> findAllById(Long id);
+public interface CommentsRepository extends JpaRepository<Comments, Integer> {
+    List<Comments> findAllById(Integer id);
 }
